+++ backends/arm/process_node.py
@@ -1,14 +1,16 @@
-# Copyright 2023-2024 Arm Limited and/or its affiliates.
+# Copyright 2024 Arm Limited and/or its affiliates.
 #
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
-
-# pyre-unsafe
+#
+from typing import cast, Dict
 
 import numpy as np
 import serializer.tosa_serializer as ts
+import torch
 import torch.fx
-from executorch.backends.arm.tosa_mapping import TosaArg
+from executorch.backends.arm.operators.node_visitor import NodeVisitor
+from executorch.backends.arm.tosa_mapping import map_dtype, TosaArg
 from executorch.backends.arm.tosa_quant_utils import (
     get_quant_arg_upstream,
     get_quantized_node_output_dtype,
@@ -16,11 +18,49 @@ from executorch.backends.arm.tosa_quant_utils import (
 )
 from executorch.backends.arm.tosa_specification import TosaSpecification
 from executorch.backends.arm.tosa_utils import (
+    getNodeArgs,
     is_bias_node_for_quantized_conv,
-    map_dtype,
     tosa_shape,
 )
 from torch.export.exported_program import ExportedProgram
+
+
+def process_call_function(
+    node: torch.fx.Node,
+    tosa_graph: ts.TosaSerializer,
+    node_visitors: Dict[str, NodeVisitor],
+    tosa_spec: TosaSpecification,
+):
+    # Unpack arguments and convert
+    inputs = getNodeArgs(node)
+
+    # Convert output (this node itself)
+    output = TosaArg(node)
+
+    is_quant_node = is_node_quantized(node)
+    if is_quant_node:
+        output_dtype = map_dtype(get_quantized_node_output_dtype(node))
+    else:
+        output_dtype = output.dtype
+    tosa_graph.currRegion.currBasicBlock.addTensor(
+        output.name,
+        tosa_shape(output.shape, output.dim_order),
+        output_dtype,
+    )
+
+    # Visiting each Node
+    # pyre-ignore[16]: Undefined attribute.
+    if node.target.__name__ in node_visitors:
+        # pyre-ignore[16]: Undefined attribute.
+        node_visitors[node.target.__name__].define_node(
+            node,
+            tosa_graph,
+            inputs,
+            output,
+            is_quant_node,
+        )
+    else:
+        raise RuntimeError(f"Unknown operator {node.target} for TOSA : {tosa_spec}")
 
 
 def process_inputs(
@@ -176,3 +216,13 @@ def process_placeholder(
         )
     else:
         raise RuntimeError(f"Placeholder '{node.name}' is of unknown type.")
+
+
+def process_output(
+    node: torch.fx.Node,
+    tosa_graph: ts.TosaSerializer,
+):
+    for output in cast(tuple[torch.fx.Node, ...], node.args[0]):
+        tosa_graph.addOutputTensor(
+            tosa_graph.currRegion.currBasicBlock.tensors[output.name]
+        )
